refactor: separate appointment and dashboard handlers
